Move from mac. Implement a basic POJO for the camera, and handle viewport in renderer
