+++ src/main/kotlin/eu/corvus/corax/graphics/Renderer.kt
@@ -1,7 +1,5 @@
 package eu.corvus.corax.graphics
 
-import eu.corvus.corax.scene.Camera
-
 /**
  * @author Vlad Ravenholm on 11/24/2019
  */
@@ -13,6 +11,4 @@ interface Renderer {
     fun onDestroy()
 
     fun onResize(width: Int, height: Int) {}
-
-    fun camera(): Camera?
 }
