+++ src/main/kotlin/eu/corvus/corax/scene/Spatial.kt
@@ -1,10 +1,7 @@
 
 package eu.corvus.corax.scene
 
-import eu.corvus.corax.graphics.Transform
 import org.joml.Matrix4f
-import org.joml.Matrix4fc
-import org.joml.Vector3f
 
 /**
  * Class that represent a positional space
@@ -15,15 +12,35 @@ import org.joml.Vector3f
  * @author Vlad Ravenholm on 11/24/2019
  */
 open class Spatial(name: String = "Spatial") : Node(name) {
-    val transform: Transform = Transform()
-    val worldTransform: Transform = Transform()
+    val transform: Transform =
+        Transform()
 
-    private val worldMatrix = Matrix4f()
+    val worldTransform: Transform =
+        Transform()
 
-    fun getWorldMatrix(): Matrix4f { //TODO cache this with the parent transforms
-        return worldMatrix.identity()
-            .translate(worldTransform.translation)
-            .rotate(worldTransform.rotation)
-            .scale(worldTransform.scale)
+    val alwaysCompute = false
+
+    val worldMatrix = Matrix4f()
+
+    private var shouldCompute = true
+
+    private fun computeWorldTransform() {
+        //TODO get parent and compute with parent
+
+        worldMatrix.identity()
+            .translate(transform.translation)
+            .rotate(transform.rotation)
+            .scale(transform.scale)
+
+        shouldCompute = false
+    }
+
+    open fun onUpdate(tpf: Float): Boolean = false
+
+    fun update(tpf: Float) {
+        val shouldUpdate = onUpdate(tpf)
+        if (alwaysCompute || shouldCompute || shouldUpdate) {
+            computeWorldTransform()
+        }
     }
 }
