+++ src/main/kotlin/eu/corvus/corax/scene/Transform.kt
@@ -27,15 +27,10 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package eu.corvus.corax.graphics
+package eu.corvus.corax.scene
 
-import eu.corvus.corax.scene.Camera
 import org.joml.Quaternionf
 import org.joml.Vector3f
-import sun.awt.windows.ThemeReader.getPosition
-import org.joml.Matrix4f
-
-
 
 /**
  * @author Vlad Ravenholm on 11/28/2019
